+++ seatunnel-engine/seatunnel-engine-storage/checkpoint-storage-api/src/main/java/org/apache/seatunnel/engine/checkpoint/storage/PipelineState.java
@@ -29,6 +29,6 @@ public class PipelineState {
 
     private String jobId;
     private int pipelineId;
-    private int checkpointId;
+    private long checkpointId;
     private byte[] states;
 }
